+++ Console/Classes/ControllerException.h
@@ -58,6 +58,6 @@
 @interface ControllerException : NSObject
 
 //convenient method to find concrete exception message by error code.
-+ (NSString *)controller:(ORControllerConfig *)controller exceptionMessageOfCode:(int)code;
++ (NSString *)controller:(ORControllerConfig *)controller exceptionMessageOfCode:(NSInteger)code;
 
 @end
