+++ Console/Classes/LogoutHelper.m
@@ -64,7 +64,7 @@
 }
 
 // Handle the server response which are from controller server with status code.
-- (void)handleServerResponseWithStatusCode:(int) statusCode {
+- (void)handleServerResponseWithStatusCode:(NSInteger) statusCode {
 	if (statusCode != 200) {
 		switch (statusCode) {
 			case UNAUTHORIZED://logout succuessful
